requirements.txt
Django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1
openai==1.25.1
python-dotenv==1.0.1
gunicorn==22.0.0
requests=2.31.0
